Temporarily mark failing test xfail (on Windows)
A strange PermissionError causes teardown of the dup_repos fixture
to fail.  This happened after we correctly switched the "remote"
to a bare repository.  This is puzzling since the teardown
explicitly handles PermissionError.  Maybe this is some other kind
of permission besides regular file attributes, maybe because the
repos are cloned in tmp_path which is in %HOME%/AppData/Local/.