workflows/test-n-publish: update python version for build step
The download wheels from PyPI step was inconsistent for 3.8 since
Spine supports Python >=3.8.1.  Remove this problem by switching to
"Build with the newest supported Python" instead of "Build with the
oldest supported Python".  This should be fine, because now the build
process is isolated, and doesn't execute any code from the package.

See:
https://github.com/spine-tools/spine-conductor/pull/31#discussion_r1740328132